Move createPncDisposalFromResult up into root of lib folder
