+++ packages/core/phase2/exceptions/HO200200.ts
@@ -3,9 +3,9 @@ import type { AnnotatedHearingOutcome } from "../../types/AnnotatedHearingOutcom
 import type Exception from "../../types/Exception"
 import type { ExceptionGenerator } from "../../types/ExceptionGenerator"
 import errorPaths from "../../lib/exceptions/errorPaths"
-import { getDisposalTextFromResult } from "../lib/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/createPncDisposalFromResult/getDisposalTextFromResult"
+import { getDisposalTextFromResult } from "../lib/createPncDisposalFromResult/getDisposalTextFromResult"
 import checkResultsMatchingPncDisposalsExceptions from "./checkResultsMatchingPncDisposalsExceptions"
-import { maxDisposalTextLength } from "../lib/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/createPncDisposalFromResult/createPncDisposalByFirstAndSecondDurations"
+import { maxDisposalTextLength } from "../lib/createPncDisposalFromResult/createPncDisposalByFirstAndSecondDurations"
 
 const generator: ExceptionGenerator = (aho: AnnotatedHearingOutcome): Exception[] => {
   const exceptions: Exception[] = []
