+++ packages/core/phase2/lib/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/isMatchToPncDisposal.ts
@@ -1,6 +1,6 @@
 import type { Result } from "../../../../types/AnnotatedHearingOutcome"
 import type { PncDisposal } from "../../../../types/PncQueryResult"
-import { createPncDisposalFromResult } from "./createPncDisposalFromResult"
+import { createPncDisposalFromResult } from "../../createPncDisposalFromResult"
 
 const isMatchToPncDisposal = (pncDisposals: PncDisposal[], result: Result): boolean =>
   createPncDisposalFromResult(result).every((ahoDisposal) =>
