+++ packages/core/phase2/lib/createPncDisposalFromResult/createPncDisposalByThirdDuration.test.ts
@@ -1,5 +1,5 @@
-import type { Result } from "../../../../../types/AnnotatedHearingOutcome"
-import DateSpecifiedInResultSequence from "../../../../../types/DateSpecifiedInResultSequence"
+import type { Result } from "../../../types/AnnotatedHearingOutcome"
+import DateSpecifiedInResultSequence from "../../../types/DateSpecifiedInResultSequence"
 import createPncDisposalByThirdDuration from "./createPncDisposalByThirdDuration"
 
 describe("createPncDisposalByThirdDuration", () => {
