+++ packages/core/phase2/lib/createPncDisposalFromResult/getDisposalTextFromResult/disqualifiedFromKeepingDisposalText.ts
