+++ packages/core/phase2/lib/createPncDisposalFromResult/getDisposalTextFromResult/getDisposalTextFromResult.test.ts
@@ -1,4 +1,4 @@
-import type { Result } from "../../../../../../types/AnnotatedHearingOutcome"
+import type { Result } from "../../../../types/AnnotatedHearingOutcome"
 import getDisposalTextFromResult from "./getDisposalTextFromResult"
 
 describe("check getDisposalTextFromResult", () => {
