+++ packages/core/phase2/lib/createPncDisposalFromResult/getDisposalTextFromResult/index.ts
