+++ packages/core/phase2/lib/createPncDisposalFromResult/getDisposalTextFromResult/isFailedToAppearWarrantIssued.ts
