+++ packages/core/phase2/lib/createPncDisposalFromResult/getDisposalTextFromResult/penaltyPointsDisposalText.test.ts
@@ -1,4 +1,4 @@
-import type { Result } from "../../../../../../types/AnnotatedHearingOutcome"
+import type { Result } from "../../../../types/AnnotatedHearingOutcome"
 import penaltyPointsDisposalText from "./penaltyPointsDisposalText"
 
 describe("check penaltyPointsDisposalText", () => {
