+++ packages/core/phase2/lib/createPncDisposalFromResult/getDisposalTextFromResult/untilFurtherOrderDisposalText.ts
