+++ packages/core/phase2/lib/createPncDisposalFromResult/getFirstDateSpecifiedInResult.test.ts
@@ -1,5 +1,5 @@
-import type { Result } from "../../../../../types/AnnotatedHearingOutcome"
-import ResultClass from "../../../../../types/ResultClass"
+import type { Result } from "../../../types/AnnotatedHearingOutcome"
+import ResultClass from "../../../types/ResultClass"
 import getFirstDateSpecifiedInResult from "./getFirstDateSpecifiedInResult"
 
 const acceptedResultClasses = [
