+++ packages/core/phase2/lib/createPncDisposalFromResult/isAmountSpecifiedInResultValid.ts
